requirements.txt
Pillow>=8.3.2
git+https://github.com/johnwmillr/LyricsGenius.git
git+https://github.com/bisoncorps/search-engine-parser
git+https://github.com/Joeclinton1/google-images-download.git
aiohttp
aria2p
async_generator
bs4
beautifulsoup4>=4.9.3
cowpy
cryptg
dnspython
emoji
gitpython
google-api-python-client
google-auth-oauthlib
googletrans==4.0.0-rc1
google-trans-new
langdetect
google_images_download>=2.8.0
gTTS>=2.1.1
gTTS-token>=1.1.3
hachoir
httplib2
humanize
heroku3
lxml==4.6.3
lyricsgenius
oauth2client
psycopg2
pybase64>=0.4.0
pylast==4.2.0
pySmartDL
python-barcode
python-dotenv
youtube-dl>=2021.1.8
youtube-search-python>=1.4.8
youtube-search>=2.1.0
pytz
qrcode
requests>=2.18.4
speedtest-cli>=2.0.2
sqlalchemy>=1.2
spotdl
selenium>=3.141.0
telethon>=1.23.0
telethon-session-sqlalchemy>=0.2.6
urbandict>=0.5
wikipedia>=1.4.0
randomstuff.py
spotify_token
twitter_scraper
chromedriver-autoinstaller
numpy
telegraph
lottie
eksipy
pydub
geopy
emoji-country-flag
cairosvg
imagedown
pyfiglet
faker